Second structural reorganization
