--- webapp/src/Wikidata/wikidataPresentation/estilo.css
@@ -1,41 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    text-align: center;
-}
-
-#quiz-container {
-    margin-top: 50px;
-}
-
-img {
-    width: 300px;
-    height: 200px;
-    object-fit: cover;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-}
-
-.option {
-    display: block;
-    margin: 10px auto;
-    padding: 10px;
-    width: 200px;
-    cursor: pointer;
-    background-color: #007BFF;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    font-size: 16px;
-    transition: background-color 0.3s;
-}
-
-.option:hover {
-    background-color: #0056b3;
-}
-
-/* 🔹 Cambia el color de los botones cuando están deshabilitados */
-.option:disabled {
-    background-color: #bdc3c7;  /* Gris */
-    color: #7f8c8d;  /* Texto más tenue */
-    cursor: not-allowed;
-}
