--- webapp/src/Wikidata/wikidataPresentation/test.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <link rel="stylesheet" href="estilo.css">
-    <title>City Quiz</title>
-</head>
-<body>
-    <script type="module" src="QuestionPresentation.js"></script>
-</body>
-</html>
